requirements.txt
pyyaml>=6.0.1
markdown>=3.5.1
tweepy>=4.14.0
